Struct invoke_function::Client

pub(crate) struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> {
    handle: Arc<Handle<C, M, R>>,
}

An ergonomic service client for AWSGirApiService.

This client allows ergonomic access to a AWSGirApiService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Fields
handle: Arc<Handle<C, M, R>>
Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Client<C, M, R>

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_layer_version_permission(&self) -> AddLayerVersionPermission<C, M, R>

Constructs a fluent builder for the AddLayerVersionPermission operation.

See AddLayerVersionPermission for more information about the operation and its arguments.

pub fn add_permission(&self) -> AddPermission<C, M, R>

Constructs a fluent builder for the AddPermission operation.

See AddPermission for more information about the operation and its arguments.

pub fn create_alias(&self) -> CreateAlias<C, M, R>

Constructs a fluent builder for the CreateAlias operation.

See CreateAlias for more information about the operation and its arguments.

pub fn create_code_signing_config(&self) -> CreateCodeSigningConfig<C, M, R>

Constructs a fluent builder for the CreateCodeSigningConfig operation.

See CreateCodeSigningConfig for more information about the operation and its arguments.

pub fn create_event_source_mapping(&self) -> CreateEventSourceMapping<C, M, R>

Constructs a fluent builder for the CreateEventSourceMapping operation.

See CreateEventSourceMapping for more information about the operation and its arguments.

pub fn create_function(&self) -> CreateFunction<C, M, R>

Constructs a fluent builder for the CreateFunction operation.

See CreateFunction for more information about the operation and its arguments.

pub fn delete_alias(&self) -> DeleteAlias<C, M, R>

Constructs a fluent builder for the DeleteAlias operation.

See DeleteAlias for more information about the operation and its arguments.

pub fn delete_code_signing_config(&self) -> DeleteCodeSigningConfig<C, M, R>

Constructs a fluent builder for the DeleteCodeSigningConfig operation.

See DeleteCodeSigningConfig for more information about the operation and its arguments.

pub fn delete_event_source_mapping(&self) -> DeleteEventSourceMapping<C, M, R>

Constructs a fluent builder for the DeleteEventSourceMapping operation.

See DeleteEventSourceMapping for more information about the operation and its arguments.

pub fn delete_function(&self) -> DeleteFunction<C, M, R>

Constructs a fluent builder for the DeleteFunction operation.

See DeleteFunction for more information about the operation and its arguments.

pub fn delete_function_code_signing_config(
&self
) -> DeleteFunctionCodeSigningConfig<C, M, R>

Constructs a fluent builder for the DeleteFunctionCodeSigningConfig operation.

See DeleteFunctionCodeSigningConfig for more information about the operation and its arguments.

pub fn delete_function_concurrency(&self) -> DeleteFunctionConcurrency<C, M, R>

Constructs a fluent builder for the DeleteFunctionConcurrency operation.

See DeleteFunctionConcurrency for more information about the operation and its arguments.

pub fn delete_function_event_invoke_config(
&self
) -> DeleteFunctionEventInvokeConfig<C, M, R>

Constructs a fluent builder for the DeleteFunctionEventInvokeConfig operation.

See DeleteFunctionEventInvokeConfig for more information about the operation and its arguments.

pub fn delete_layer_version(&self) -> DeleteLayerVersion<C, M, R>

Constructs a fluent builder for the DeleteLayerVersion operation.

See DeleteLayerVersion for more information about the operation and its arguments.

pub fn delete_provisioned_concurrency_config(
&self
) -> DeleteProvisionedConcurrencyConfig<C, M, R>

Constructs a fluent builder for the DeleteProvisionedConcurrencyConfig operation.

See DeleteProvisionedConcurrencyConfig for more information about the operation and its arguments.

pub fn get_account_settings(&self) -> GetAccountSettings<C, M, R>

Constructs a fluent builder for the GetAccountSettings operation.

See GetAccountSettings for more information about the operation and its arguments.

pub fn get_alias(&self) -> GetAlias<C, M, R>

Constructs a fluent builder for the GetAlias operation.

See GetAlias for more information about the operation and its arguments.

pub fn get_code_signing_config(&self) -> GetCodeSigningConfig<C, M, R>

Constructs a fluent builder for the GetCodeSigningConfig operation.

See GetCodeSigningConfig for more information about the operation and its arguments.

pub fn get_event_source_mapping(&self) -> GetEventSourceMapping<C, M, R>

Constructs a fluent builder for the GetEventSourceMapping operation.

See GetEventSourceMapping for more information about the operation and its arguments.

pub fn get_function(&self) -> GetFunction<C, M, R>

Constructs a fluent builder for the GetFunction operation.

See GetFunction for more information about the operation and its arguments.

pub fn get_function_code_signing_config(
&self
) -> GetFunctionCodeSigningConfig<C, M, R>

Constructs a fluent builder for the GetFunctionCodeSigningConfig operation.

See GetFunctionCodeSigningConfig for more information about the operation and its arguments.

pub fn get_function_concurrency(&self) -> GetFunctionConcurrency<C, M, R>

Constructs a fluent builder for the GetFunctionConcurrency operation.

See GetFunctionConcurrency for more information about the operation and its arguments.

pub fn get_function_configuration(&self) -> GetFunctionConfiguration<C, M, R>

Constructs a fluent builder for the GetFunctionConfiguration operation.

See GetFunctionConfiguration for more information about the operation and its arguments.

pub fn get_function_event_invoke_config(
&self
) -> GetFunctionEventInvokeConfig<C, M, R>

Constructs a fluent builder for the GetFunctionEventInvokeConfig operation.

See GetFunctionEventInvokeConfig for more information about the operation and its arguments.

pub fn get_layer_version(&self) -> GetLayerVersion<C, M, R>

Constructs a fluent builder for the GetLayerVersion operation.

See GetLayerVersion for more information about the operation and its arguments.

pub fn get_layer_version_by_arn(&self) -> GetLayerVersionByArn<C, M, R>

Constructs a fluent builder for the GetLayerVersionByArn operation.

See GetLayerVersionByArn for more information about the operation and its arguments.

pub fn get_layer_version_policy(&self) -> GetLayerVersionPolicy<C, M, R>

Constructs a fluent builder for the GetLayerVersionPolicy operation.

See GetLayerVersionPolicy for more information about the operation and its arguments.

pub fn get_policy(&self) -> GetPolicy<C, M, R>

Constructs a fluent builder for the GetPolicy operation.

See GetPolicy for more information about the operation and its arguments.

pub fn get_provisioned_concurrency_config(
&self
) -> GetProvisionedConcurrencyConfig<C, M, R>

Constructs a fluent builder for the GetProvisionedConcurrencyConfig operation.

See GetProvisionedConcurrencyConfig for more information about the operation and its arguments.

pub fn invoke(&self) -> Invoke<C, M, R>

Constructs a fluent builder for the Invoke operation.

See Invoke for more information about the operation and its arguments.

pub fn invoke_async(&self) -> InvokeAsync<C, M, R>

Constructs a fluent builder for the InvokeAsync operation.

See InvokeAsync for more information about the operation and its arguments.

pub fn list_aliases(&self) -> ListAliases<C, M, R>

Constructs a fluent builder for the ListAliases operation.

See ListAliases for more information about the operation and its arguments.

pub fn list_code_signing_configs(&self) -> ListCodeSigningConfigs<C, M, R>

Constructs a fluent builder for the ListCodeSigningConfigs operation.

See ListCodeSigningConfigs for more information about the operation and its arguments.

pub fn list_event_source_mappings(&self) -> ListEventSourceMappings<C, M, R>

Constructs a fluent builder for the ListEventSourceMappings operation.

See ListEventSourceMappings for more information about the operation and its arguments.

pub fn list_function_event_invoke_configs(
&self
) -> ListFunctionEventInvokeConfigs<C, M, R>

Constructs a fluent builder for the ListFunctionEventInvokeConfigs operation.

See ListFunctionEventInvokeConfigs for more information about the operation and its arguments.

pub fn list_functions(&self) -> ListFunctions<C, M, R>

Constructs a fluent builder for the ListFunctions operation.

See ListFunctions for more information about the operation and its arguments.

pub fn list_functions_by_code_signing_config(
&self
) -> ListFunctionsByCodeSigningConfig<C, M, R>

Constructs a fluent builder for the ListFunctionsByCodeSigningConfig operation.

See ListFunctionsByCodeSigningConfig for more information about the operation and its arguments.

pub fn list_layers(&self) -> ListLayers<C, M, R>

Constructs a fluent builder for the ListLayers operation.

See ListLayers for more information about the operation and its arguments.

pub fn list_layer_versions(&self) -> ListLayerVersions<C, M, R>

Constructs a fluent builder for the ListLayerVersions operation.

See ListLayerVersions for more information about the operation and its arguments.

pub fn list_provisioned_concurrency_configs(
&self
) -> ListProvisionedConcurrencyConfigs<C, M, R>

Constructs a fluent builder for the ListProvisionedConcurrencyConfigs operation.

See ListProvisionedConcurrencyConfigs for more information about the operation and its arguments.

pub fn list_tags(&self) -> ListTags<C, M, R>

Constructs a fluent builder for the ListTags operation.

See ListTags for more information about the operation and its arguments.

pub fn list_versions_by_function(&self) -> ListVersionsByFunction<C, M, R>

Constructs a fluent builder for the ListVersionsByFunction operation.

See ListVersionsByFunction for more information about the operation and its arguments.

pub fn publish_layer_version(&self) -> PublishLayerVersion<C, M, R>

Constructs a fluent builder for the PublishLayerVersion operation.

See PublishLayerVersion for more information about the operation and its arguments.

pub fn publish_version(&self) -> PublishVersion<C, M, R>

Constructs a fluent builder for the PublishVersion operation.

See PublishVersion for more information about the operation and its arguments.

pub fn put_function_code_signing_config(
&self
) -> PutFunctionCodeSigningConfig<C, M, R>

Constructs a fluent builder for the PutFunctionCodeSigningConfig operation.

See PutFunctionCodeSigningConfig for more information about the operation and its arguments.

pub fn put_function_concurrency(&self) -> PutFunctionConcurrency<C, M, R>

Constructs a fluent builder for the PutFunctionConcurrency operation.

See PutFunctionConcurrency for more information about the operation and its arguments.

pub fn put_function_event_invoke_config(
&self
) -> PutFunctionEventInvokeConfig<C, M, R>

Constructs a fluent builder for the PutFunctionEventInvokeConfig operation.

See PutFunctionEventInvokeConfig for more information about the operation and its arguments.

pub fn put_provisioned_concurrency_config(
&self
) -> PutProvisionedConcurrencyConfig<C, M, R>

Constructs a fluent builder for the PutProvisionedConcurrencyConfig operation.

See PutProvisionedConcurrencyConfig for more information about the operation and its arguments.

pub fn remove_layer_version_permission(
&self
) -> RemoveLayerVersionPermission<C, M, R>

Constructs a fluent builder for the RemoveLayerVersionPermission operation.

See RemoveLayerVersionPermission for more information about the operation and its arguments.

pub fn remove_permission(&self) -> RemovePermission<C, M, R>

Constructs a fluent builder for the RemovePermission operation.

See RemovePermission for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_alias(&self) -> UpdateAlias<C, M, R>

Constructs a fluent builder for the UpdateAlias operation.

See UpdateAlias for more information about the operation and its arguments.

pub fn update_code_signing_config(&self) -> UpdateCodeSigningConfig<C, M, R>

Constructs a fluent builder for the UpdateCodeSigningConfig operation.

See UpdateCodeSigningConfig for more information about the operation and its arguments.

pub fn update_event_source_mapping(&self) -> UpdateEventSourceMapping<C, M, R>

Constructs a fluent builder for the UpdateEventSourceMapping operation.

See UpdateEventSourceMapping for more information about the operation and its arguments.

pub fn update_function_code(&self) -> UpdateFunctionCode<C, M, R>

Constructs a fluent builder for the UpdateFunctionCode operation.

See UpdateFunctionCode for more information about the operation and its arguments.

pub fn update_function_configuration(
&self
) -> UpdateFunctionConfiguration<C, M, R>

Constructs a fluent builder for the UpdateFunctionConfiguration operation.

See UpdateFunctionConfiguration for more information about the operation and its arguments.

pub fn update_function_event_invoke_config(
&self
) -> UpdateFunctionEventInvokeConfig<C, M, R>

Constructs a fluent builder for the UpdateFunctionEventInvokeConfig operation.

See UpdateFunctionEventInvokeConfig for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(
conf: Config,
conn: C
) -> Client<C, AwsMiddleware, Standard>

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from the service Config.